operate health display checks on the invisible flag
I changed from invincible to invisible. Looks like a typo, doesn't it? But no.
It's just a simpler, more accurate way of determining who needs a health bar.
Although it introduces an issue with the Picopico blocks that can't be fixed
with a whitelist or blacklist, described in the comments.